libs/shared/src/lib/switch/switch.component.ts
import { Component, Input, HostBinding, forwardRef } from '@angular/core';
import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';

@Component({
  selector: 'cm-switch',
  templateUrl: './switch.component.html',
  styleUrls: ['./switch.component.scss'],
  providers: [
    {
      provide: NG_VALUE_ACCESSOR,
      useExisting: forwardRef(() => SwitchComponent),
      multi: true,
    },
  ],
})
export class SwitchComponent implements ControlValueAccessor {
  @Input() label?: string;

  disabled = false;
  on = false;

  @HostBinding('class.cm-switch')
  get defaultClass() {
    return true;
  }

  toggleSwitch() {
    this.on = !this.on;
    this.onChange(this.on);
    this.onTouched();
  }

  onChange: any = () => {};
  onTouched: any = () => {};

  registerOnChange(fn: () => any) {
    this.onChange = fn;
  }

  writeValue(value: boolean) {
    this.on = !!value;
  }

  registerOnTouched(fn: () => any) {
    this.onTouched = fn;
  }

  setDisabledState(isDisabled: boolean): void {
    this.disabled = isDisabled;
  }
}
